refactor : 구조/클래스 명 변경
