+++ src/main/java/nextstep/RoomEscapeSpringApplication.java
@@ -1,4 +1,4 @@
-package nextstep.reservation;
+package nextstep;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
