+++ src/main/java/nextstep/reservation/ReservationRepository.java
@@ -1,5 +1,7 @@
 package nextstep.reservation;
 
+import nextstep.reservation.domain.Reservation;
+
 import java.time.LocalDate;
 import java.time.LocalTime;
