+++ src/main/java/nextstep/reservation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package nextstep.reservation;
+package nextstep.reservation.domain;
 
 import org.springframework.lang.Nullable;
